fix(e2e): Remove hardcoded path in universal_client_test.go
The `universal_client_test.go` was using a hardcoded path to the project root, which caused the e2e tests to fail in the GHA environment.

This commit removes the hardcoded path, allowing the `unimock` binary to be executed from the current working directory, which is the project root in the GHA environment.